requirements.txt
#for streamlit
streamlit==1.2.0
streamlit-webrtc==0.31.3
av

#for Deep learning model
tensorflow==2.6.1

#Image
Pillow

#other
matplotlib
numpy

